Swich the Type of the list items in MediaInfo from Double to Long

+++ VLCBenchmark/app/src/main/java/org/videolan/vlcbenchmark/service/MediaInfo.java
@@ -8,11 +8,11 @@ import java.util.List;
 public class MediaInfo {
         String url;
         String name;
-        List<Double> snapshot;
+        List<Long> snapshot;
         String checksum;
         String localUrl;
 
-    public MediaInfo(String url, String name, String checksum, List<Double> snapshot) {
+    public MediaInfo(String url, String name, String checksum, List<Long> snapshot) {
         this.url = url;
         this.name = name;
         this.checksum = checksum;
